Panner GUI: fix detent in center
Previously the dead-zone was too small (1/360). The mono/balance
panner GUI has a throw of 180 deg L<>R.

Also snapping to center didn't allow to smoothly move out of the
center. The accumulated_delta as directly applied. This caused
jumps by 4.5. degrees.

This commit reduces the deadzone to 1 degree of the azimuth
around the center.