Ripple move simplifications
Dropping the feature of rippling across tracks makes things simpler and allows
removal of a quite a few lines of code.

As we no longer ripple move regions from one track to another, we no longer
need to keep track which regions are selected during the drag and which regions
are moved due to rippling. We can treat all regions after the first selected
one in one track as if they were selected as well.

This also fixes properly what was already addressed in 9b2228d6fd.
Unfortunately that did only fix the behavior when ripple moving
forwards. This fix also works ripple moving backwards.