Editor: replicate nudge-forward/nudge-backward in the Editor/ action namespace and use these actions for the default key bindings
Previously the bindings were given to the Region/* equivalents and made insensitive if there was no selected region.

Given that we have specific actions for nudge-playhead-*, and the ones in Region/ will always work on regions
if there are any selected regions, this provides a better set of possible actions, even if the naming is not
totally consistent.