Increase timeout for export-preparation
In some rare cases the butler may be busy for a long time
(e.g. seek in a large session may still be active when
export is started).

8 second timeout seems reasonable to prevent the app from hanging
without desktops showing a "unresponsive" popup in case the
timeout is reached for some reason (e.g the engine dies, and
no TFSM transition can happen).