Update mixer-strip group button height
On tracks the bottom row of each mixer-strip has three buttons.
Two are provided by the gain-meter (automation, meter-point),
and the size of those is set in GainMeterBase (current 15x15px).

The track-group button size was left unspecified.
The master-bus only has the gain-meter buttons, and hence was
less tall compared to tracks.