maximized region mode
When the user makes "Ctrl + double click" on the region
the region view will be maximized and the visual state just before this will be kept.
This will make like entering into a maximized region mode.
The next "Ctrl + double click" will return visual state to the kept one,
i.e. leaving the maximized region mode.
When the visual state is in maximized mode and there vertical scroll is done
than the mode will be just cleared and the visual state will not be changed.