Revert "Fix auto capture alignment when bouncing metronome"
When the metronome is routed through a track, it
will correctly get the downstream playback latency.
Capture Alignment is Capture-Time.

This fixes a case where the metronome is connected to
physical I/O and a track. This can result in ambiguous
latency if the track's output is not connected or
connected via latent master bus.

This reverts commit f08b90f36870249cc3e9ea089dd6594269c67148.