WIP: Fix namespace issues and build with GCC8
The build was broken for me with both GCC 8 and clang 11 due to the lookup of
operator<<().  However, since the previous pattern of using a namespace then
defining things in that namespace in... the global namespace... sort of... is
very strange, and likely to cause further problems with ADL especially as we
move to newer language versions and libraries, I opted to go all-out here and
define things inside the appropriate namespace.

This will probably resolve some earlier issues with clang and MSVC as well,
since they each use different lookup rules that all have their own quirks
around this stuff.