Define operator<< in the corresponding namespace
This fixes the build for me with clang 11.1.0 (from LLVM repositories) on
Debian buster as C++11.  As far as I know this is the general best practice
because of ADL voodoo, but Ardour's approach to namespaces is pretty anarchic,
and I'm not really sure exactly why this wasn't building for some types.

Also tested to build find with the default buster GCC 8.3.0.