WIP: Add SessionControllerHandleRef
This can be used to ensure that whenever this has a reference to a session, it
also has a session controller set up to use that session.  Towards sharing code
between ARDOUR_UI and SessionController.

WIP: SessionHandleRef|Ptr are not safe abstractions, so this isn't either,
since derived types can and do manually mess around with the session pointer.
Probably lifetime bugs here, and probably a good idea to statically prevent
that, but a /lot/ of existing code uses _session directly.