VST3: prefer linear knob movement
The previous idea was to follow Ardour's default
 \delta val = \delta Y - \delta X

In reality it turns out that this is better represented
with kLinearMode.

VST3 offers the following:
  * Circular with jump to clicked position
  * Circular without jump to clicked position
  * Linear: depending on vertical movement

The first is right out. Value changes need to be
continuous. The last explicitly mentioning vertical
movement was also somewhat discouraging,
however most plugins don't take it literally and allow
linear X/Y movement.